+++ public/js/index.js
@@ -1,10 +1,10 @@
 // Get references to page elements
-require("dotenv").config();
+
 var $exampleText = $("#example-text");
 var $exampleDescription = $("#example-description");
 var $submitBtn = $("#submit");
 var $exampleList = $("#example-list");
-var api_key = process.env.API_KEY;
+
 
 // The API object contains methods for each kind of request we'll make
 var API = {
@@ -110,38 +110,7 @@ $("#find-movie").on("click", function (event) {
     // Here we grab the text from the input box
     var movie = $("#movie-input").val();
 
-    // UTelly config
-    var settings = {
-        "async": true,
-        "crossDomain": true,
-        "url": "https://utelly-tv-shows-and-movies-availability-v1.p.rapidapi.com/lookup?term=" + movie + "&country=us",
-        "method": "GET",
-        "headers": {
-            "x-rapidapi-host": "utelly-tv-shows-and-movies-availability-v1.p.rapidapi.com",
-            "x-rapidapi-key": api_key
-        }
-    }
-
-    //UTelly Call
-    $.ajax(settings).done(function (response) {
-        var res = response.results;
-        console.log(res);
-        for (var i = 0; i < res.length; i++) {
-            $("#movie-view").append('<div class="movie-data pt-5"><h3>' +
-                res[i].name + '</h3><br>' +
-                '<img class="movie-pic img-fluid" src=' + res[i].picture + '><br></div>');
-            for (var j = 0; j < res[i].locations.length; j++) {
-                if (res[i].locations[j].icon) {
-                    $("#movie-view").append(
-                         '<div class="streaming-list"><a target="_blank" href=' + res[i].locations[j].url + 
-                         '><img class="streaming-icons img-fluid" src=' + res[i].locations[j].icon + '></a></div>'
-                    );
-                }
-            }
-        }
-
-    });
-
+    
     // OMDB
     var queryURL = "https://www.omdbapi.com/?t=" + movie + "&apikey=trilogy";
